se movio el handle upload file TDD log a la logica de la aplicacion de los assignments, por arquitectura y por que vamos a reutilizarlo
