+++ Web Ui/src/modules/Assignments/application/HandleUploadTDDLogFile.ts
@@ -1,0 +1,67 @@
+import JSZip from "jszip";
+import CryptoJS from "crypto-js";
+import { VITE_API } from "../../../../config";
+
+export const handleUploadTDDLogFile = async (file: File, repositoryLink?: string): Promise<void> => {
+  try {
+    console.log("Archivo recibido:", file);
+    const reader = new FileReader();
+
+    reader.onload = async () => {
+      const encryptedContent = reader.result as string;
+      const binaryData = decryptContent(encryptedContent, "iDHJp8o32$%u4drMjPLq8c!7S@wZEXWC");
+      const fileContent = await extractFileFromZip(binaryData, "tdd_log.json");
+      const jsonData = parseJSON(fileContent);
+
+      const updatedData = enrichWithRepoData(jsonData, repositoryLink);
+      console.log("JSON actualizado:", updatedData);
+      const response = await fetch(`${VITE_API}/TDDCycles/upload-log`, {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify(updatedData),
+      });
+
+      if (!response.ok) {
+        throw new Error(`Error en el POST: ${response.statusText}`);
+      }
+
+      const responseData = await response.json();
+      console.log("Respuesta de la API:", responseData);
+    };
+
+    reader.readAsText(file);
+  } catch (error) {
+    console.error("Error al procesar el archivo:", error);
+    throw error;
+  }
+};
+
+const decryptContent = (encryptedContent: string, decryptionKey: string): Uint8Array => {
+  const decryptedBytes = CryptoJS.AES.decrypt(encryptedContent, decryptionKey);
+  const base64Data = decryptedBytes.toString(CryptoJS.enc.Utf8);
+  if (!base64Data) throw new Error("Error al desencriptar");
+  return Uint8Array.from(atob(base64Data), (char) => char.charCodeAt(0));
+};
+
+const extractFileFromZip = async (binaryData: Uint8Array, targetFileName: string): Promise<string> => {
+  const zip = new JSZip();
+  const loadedZip = await zip.loadAsync(binaryData);
+  const targetFile = loadedZip.file(targetFileName);
+  if (!targetFile) throw new Error(`Archivo ${targetFileName} no encontrado en el ZIP.`);
+  return targetFile.async("string");
+};
+
+const parseJSON = (fileContent: string): any => {
+  try {
+    return JSON.parse(fileContent);
+  } catch {
+    throw new Error("Error al parsear el JSON.");
+  }
+};
+
+const enrichWithRepoData = (jsonData: any, repoLink?: string) => {
+  if (!repoLink) throw new Error("Enlace del repositorio no proporcionado.");
+  const repoMatch = repoLink.match(/https:\/\/github\.com\/([^/]+)\/([^/]+)/);
+  if (!repoMatch) throw new Error("Enlace del repositorio inválido.");
+  return { repoName: repoMatch[2], repoOwner: repoMatch[1], log: jsonData };
+};
